hotel_compare/requirements.txt
# 必須パッケージ
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.17.0

# オプショナル（fuzzy matching用）
rapidfuzz>=3.0.0

# オプショナル（NLP用、重いので必要に応じてインストール）
# transformers>=4.30.0
# torch>=2.0.0

